group the test files
